segragate story book from wallet root page
